Create main.py
A game of war that simulates a deck of cards and their respective values, a table to play on, the drawing of random cards from the player's decks and the occurrence of when a war condition is met, i.e, both players draw the same card. See the read me for more details!